Program works
