+++ src/main/java/mozeik/gerrod/CurrencyConverter.java
@@ -7,21 +7,16 @@ import java.math.RoundingMode;
  * Created by gerrodmozeik on 1/12/17.
  */
 public class CurrencyConverter {
-    Double relativeExchangeRate;
-    Double inputMonetaryValue;
-    BigDecimal bd = new BigDecimal(0);
+    double relativeExchangeRate;
+    double inputMonetaryValue;
     double roundedOutputValue = 0;
 
-    public static double round(double value, int places) {
-        BigDecimal bd = new BigDecimal(value);
-        bd = bd.setScale(places, RoundingMode.HALF_UP);
-        return bd.doubleValue();
-    }
 
     public Double convertInputValueToOutputValue(Double relativeExchangeRate, Double inputMonetaryValue) {
+        CurrencyRounder currencyRounder = new CurrencyRounder();
         this.relativeExchangeRate = relativeExchangeRate;
         this.inputMonetaryValue = inputMonetaryValue;
-        this.roundedOutputValue = round(relativeExchangeRate * inputMonetaryValue, 2);
+        this.roundedOutputValue = currencyRounder.round(relativeExchangeRate * inputMonetaryValue, 2);
         return roundedOutputValue;
     }
 }
