+++ src/main/java/mozeik/gerrod/Prompter.java
@@ -1,14 +1,19 @@
 package mozeik.gerrod;
 
+import java.util.Date;
 import java.util.Scanner;
 
 /**
  * Created by gerrodmozeik on 1/13/17.
  */
 public class Prompter {
+    char caseIdentifier;
+    Double inputMonetaryValue;
+    Double outputMonetaryValue;
+
     Scanner scanner = new Scanner(System.in);
 
-    public char caseIdentifier() {
+    public char getCaseIdentifier() {
         System.out.println("What type of conversion would you like?");
         System.out.println("a: Dollar to Euro");
         System.out.println("b: Euro to Dollar");
@@ -19,12 +24,20 @@ public class Prompter {
         System.out.println("g: Singapore Dollar to Swiss Franc");
         System.out.println("h: Swiss Franc to Malaysian Ringgit");
         System.out.println("i: Malaysian Ringgit to Japanese Yen");
-        System.out.println("j: Japanese Yen to Chinese Yuan");
-        return scanner.next().charAt(0);
+        System.out.println("j: Japanese Yen to Chinese Yuan Renminbi");
+        caseIdentifier = scanner.next().charAt(0);
+        System.out.println(caseIdentifier);
+        return caseIdentifier;
     }
 
-    public Double inputMonetaryValue() {
+    public Double getInputMonetaryValue() {
         System.out.println("How much money would you like to convert?");
-        return scanner.nextDouble();
+        inputMonetaryValue = scanner.nextDouble();
+        System.out.println(inputMonetaryValue);
+        return inputMonetaryValue;
+    }
+
+    public void displayConvertedMonetaryValue(Double outputMonetaryValue) {
+        System.out.println("Your converted amount is: " + outputMonetaryValue);
     }
 }
