--- src/main/java/mozeik/gerrod/main.java
@@ -1,21 +1,0 @@
-package mozeik.gerrod;
-
-import java.util.Scanner;
-
-/**
- * Created by gerrodmozeik on 1/12/17.
- */
-/*public class main {
-
-
-    public static void main(String[] args) {
-        WuTangFinancial w = new WuTangFinancial();
-        System.out.println("Which case are you referencing?");
-        Scanner scanner = new Scanner(System.in);
-        char caseIdentifier = scanner.next();
-        System.out.println("How much do you want to convert?");
-        Double inputMonetaryValue = scanner.nextDouble();
-        w.run(caseIdentifier, inputMonetaryValue);
-    }
-}
-*/
